Spear buff ## About The Pull Request
I've made a severe and terrible lapse in judgement, and am reverting most of the overzealous polearm nerfs. They should be usable again. More fine tuning to come, but this PR just fixes a solution that was worse than the problem. I will be more careful with balancing in the future and extend my heartfelt apologies to the spear mafia.
## Testing Evidence

<!-- It's mandatory to test your PR. Provide images, clips or description of how you tested your changes where possible. -->

## Why It's Good For The Game

Spears should be usable.

<!-- By contributing to this codebase, you confirm that any code and sprites you provide are legal to share and will be licensed under the terms specified in README.md — AGPLv3 for code and CC-BY-SA 3.0 for assets, unless otherwise stated. You acknowledge that the project maintainers are under no obligation to remove any materials that do not violate these licenses. -->