Adds big shark tail to shark race (and also redirect the github link in config to our current codebase) (#237)
* Adds big shark tail to shark race

It was requested to add it to them, and it makes sense for them to have the option too. Give the sharks fat shark tails!

* Redirect Github Address in config to current address

Don't know why we didn't do it for this long. Just changes the config to go to the current version of Caustic Cove instead of the old one.